Move accounts files to separate package
